\docType{package}
\name{IgorR-package}
\alias{IgorR}
\alias{IgorR-package}
\title{Read binary files saved by Igor Pro (including Neuromatic data)}
\description{
This package provides function to read data from the Igor
Pro data analysis program by Wavemetrics. The data formats
supported are Igor packed experiment format (pxp) and Igor
binary wave (ibw). See: http://www.wavemetrics.com/ for
details.
}
\details{
It also includes some functions to load special pxp files
produced by the Neuromatic/Nclamp packages for recording
and analysing neuronal #data. See
http://www.neuromatic.thinkrandom.com/ for details.

Key functions in the package include
\code{\link{read.ibw},\link{read.pxp}}.
}
\keyword{package}

