% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIgorBinary.R
\name{WaveToTimeSeries}
\alias{WaveToTimeSeries}
\title{Convert an Igor wave (wave list) loaded by read.ibw into an R time series}
\usage{
WaveToTimeSeries(WaveData, ReturnOriginalDataOnError = TRUE)
}
\arguments{
\item{WaveData, }{a wave or list of waves}

\item{ReturnOriginalDataOnError}{If we can't make a time series, return
return original data (default TRUE)}
}
\value{
a time series or multi time series (\code{\link[stats]{ts}},
  \code{mts})
}
\description{
Where there are multiple waves, they are assumed to be of compatible lengths
so that they can be joined together by \code{cbind}.
}
\seealso{
Other igor-io: 
\code{\link{read.ibw}()},
\code{\link{read.pxp}()}
}
\author{
jefferis
}
\concept{igor-io}
