% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadNclamp.R
\name{ReadAllNclampLogTables}
\alias{ReadAllNclampLogTables}
\title{Read all Nclamp log tables from a directory into a list}
\usage{
ReadAllNclampLogTables(logfiledir, pattern = "_log[0-9]+[.]pxp$", ...)
}
\arguments{
\item{logfiledir}{path to directory containing log files (pxp files).}

\item{pattern}{optional regular expression -- see \code{\link{list.files}}.}

\item{...}{additional parameters for \code{\link{ReadNclampLogTable}}.}
}
\value{
A named list containing one dataframe for each parsed log file.
}
\description{
Read all Nclamp log tables from a directory into a list
}
\examples{
\dontrun{
logfiles <- ReadAllNclampLogTables("/GD/projects/PhysiologyData/logs")
str(logfiles)
}
}
\author{
jefferis
}
