% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadNclamp.R
\name{ReadNclampLogTable}
\alias{ReadNclampLogTable}
\title{Read the log table produced by Nclamp acquisition software for Igor}
\usage{
ReadNclampLogTable(f, Verbose = FALSE)
}
\arguments{
\item{f}{path to the log file.}

\item{Verbose}{whether to print status information while reading the file.}
}
\value{
A dataframe containing a row for each acquisition protocol run.
}
\description{
log tables are special Igor .pxp files that contain only variables.
Each entry corresponds to a single run of an Nclamp protocol, 
storing information like protocol name, run time etc.
}
\author{
jefferis
}
