% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIgorBinary.R
\name{tsp.igorwave}
\alias{tsp.igorwave}
\title{Return tsp attribute of Igor wave (start, end, frequency)}
\usage{
tsp.igorwave(wave)
}
\arguments{
\item{wave}{Igor wave loaded by \code{read.ibw} or\code{read.pxp}}
}
\value{
numeric vector with elements start, end, frequency
}
\description{
Note that end = (npts-1) * deltat
}
\seealso{
\code{\link[stats]{tsp}}
}
\author{
jefferis
}
